# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "vm okr 2023",
        "description": "VM ORK 2023",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "local server"
        }
    ],
    "paths": {
        "/api/orders/add": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Add a new order",
                "description": "Example request:\n\n```\ncurl -X POST \"http://localhost/api/orders/add\" -H \"Content-Type: application/x-www-form-urlencoded\" -d \"name=Dr. Mr. Lakatos Marlonbrando Winettou\" -d \"email=winettou@lakatosestarsa.hu\" -d \"receipt=home_delivery\" -d \"billing_name=Lakatos Marlonbrando Winettou\" -d \"billing_postal_code=9999\" -d \"billing_city=Kukutyin\" -d \"billing_public_space=Kultúra útja\" -d \"billing_house_number=12\" -d \"delivery_name=Lakatos Marlonbrando Winettou\" -d \"delivery_postal_code=6666\" -d \"delivery_city=Piripócs\" -d \"delivery_public_space=Alkoholista tér\" -d \"delivery_house_number=32/a\" -d \"products[0][name]=Latex búslakodó\" -d \"products[0][quantity]=2\" -d \"products[0][actual_price]=25599\" -d \"products[1][name]=Sertés taraj hegyező\" -d \"products[1][quantity]=1\" -d \"products[1][actual_price]=6599.99\"\n```",
                "operationId": "34007df1e27e183411c075607e47c144",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "receipt",
                                    "billing_name",
                                    "billing_postal_code",
                                    "billing_city",
                                    "billing_public_space",
                                    "billing_house_number",
                                    "delivery_name",
                                    "delivery_postal_code",
                                    "delivery_city",
                                    "delivery_public_space",
                                    "delivery_house_number",
                                    "products"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string"
                                    },
                                    "receipt": {
                                        "description": "receipt",
                                        "type": "string"
                                    },
                                    "billing_name": {
                                        "description": "billing name",
                                        "type": "string"
                                    },
                                    "billing_postal_code": {
                                        "description": "billing postal code",
                                        "type": "string"
                                    },
                                    "billing_city": {
                                        "description": "billing city",
                                        "type": "string"
                                    },
                                    "billing_public_space": {
                                        "description": "billing public space",
                                        "type": "string"
                                    },
                                    "billing_house_number": {
                                        "description": "billing house number",
                                        "type": "string"
                                    },
                                    "delivery_name": {
                                        "description": "delivery name",
                                        "type": "string"
                                    },
                                    "delivery_postal_code": {
                                        "description": "delivery postal code",
                                        "type": "string"
                                    },
                                    "delivery_city": {
                                        "description": "delivery city",
                                        "type": "string"
                                    },
                                    "delivery_public_space": {
                                        "description": "delivery public space",
                                        "type": "string"
                                    },
                                    "delivery_house_number": {
                                        "description": "delivery house number",
                                        "type": "string"
                                    },
                                    "products": {
                                        "description": "products",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "description": "product name",
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "description": "product quantity",
                                                    "type": "integer"
                                                },
                                                "actual_price": {
                                                    "description": "actual price of the product",
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"orderId\": 123\n                }"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Bad Request\",\n                    \"message\": \"Invalid input data\"\n                }"
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Server Error\",\n                    \"message\": \"Internal server error occurred\"\n                }"
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/list": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "description": "Example request:\n\n```\ncurl -X POST \"http://localhost/api/orders/list\" -H \"Content-Type: application/x-www-form-urlencoded\" -d \"orderId=123\"\n```",
                "operationId": "8a09b12055b683fcd65fa36e16d9d86d",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "orderId": {
                                        "description": "ID of the order to retrieve",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "Status of the orders to retrieve",
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "completed"
                                        ]
                                    },
                                    "startDate": {
                                        "description": "Start date of the orders to retrieve",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "endDate": {
                                        "description": "End date of the orders to retrieve",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "example": "\n                    {\n                        \"orders\": [\n                            {\n                                \"id\": 123,\n                                \"status\" : \"new\",\n                                \"created_at\": \"2024-03-19T12:07:53.000000Z\",\n                                \"name\": \"Dr. Mr. Lakatos Marlonbrando Winettou\",\n                                \"total_amount\": 57797.99\n                            }\n                        ]\n                    }"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Bad Request\",\n                    \"message\": \"Invalid input data\"\n                }"
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Server Error\",\n                    \"message\": \"Internal server error occurred\"\n                }"
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/modify": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Modify an existing order status",
                "description": "Example request:\n\n```\ncurl -X POST \"http://localhost/api/orders/modify\" -H \"Content-Type: application/x-www-form-urlencoded\" -d \"orderId=123\" -d \"newStatus=completed\"\n```",
                "operationId": "3bf9b4c11a608497e0b709fe12c326c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "orderId",
                                    "newStatus"
                                ],
                                "properties": {
                                    "orderId": {
                                        "description": "The ID of the modified order",
                                        "type": "integer"
                                    },
                                    "newStatus": {
                                        "description": "The new status of the order",
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "example": "\n                    [\n                        123,\n                        \"completed\"\n                    ]"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Not Found\",\n                    \"message\": \"Order not found\"\n                }"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Bad Request\",\n                    \"message\": \"Invalid input data\"\n                }"
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "example": "\n                {\n                    \"error\": \"Server Error\",\n                    \"message\": \"Internal server error occurred\"\n                }"
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}